+++ src/api/nft.ts
@@ -6,6 +6,7 @@ import { Media, NftMetadata, NftTokenType, TokenUri } from '../types/types';
  * @public
  */
 export interface BaseNftContract {
+  /** The address of the contract. */
   address: string;
 }
 
@@ -44,6 +45,9 @@ export interface BaseNft {
  * @public
  */
 export interface Nft extends BaseNft {
+  /** The NFT's underlying contract and relevant contract metadata. */
+  contract: NftContract;
+
   /** The NFT title. */
   title: string;