+++ src/api/transact-namespace.ts
@@ -6,7 +6,16 @@ import {
   TransactionResponse
 } from '@ethersproject/abstract-provider';
 
+/**
+ * The Transact namespace contains methods used for sending transactions and
+ * checking on the state of submitted transactions.
+ *
+ * Do not call this constructor directly. Instead, instantiate an Alchemy object
+ * with `const alchemy = new Alchemy(config)` and then access the core namespace
+ * via `alchemy.transact`.
+ */
 export class TransactNamespace {
+  /** @internal */
   constructor(private readonly config: AlchemyConfig) {}
 
   /**
@@ -16,8 +25,8 @@ export class TransactNamespace {
    * Returns the transaction hash of the submitted transaction.
    *
    * @param signedTransaction The raw, signed transaction as a hash.
-   * @param maxBlockNumber Optional hex-encoded number string. Highest block
-   *   number in which the transaction should be included.
+   * @param maxBlockNumber Optional highest block number in which the
+   *   transaction should be included.
    * @param options Options to configure the request.
    */
   async sendPrivateTransaction(